Text.Libyaml

Description

Low-level, streaming YAML interface. For a higher-level interface, see Data.Yaml.

The event stream

data MarkedEvent

Event with start and end marks.

Since: 0.10.4.0

Constructors

MarkedEvent

Fields

yamlEvent :: Event

yamlStartMark :: YamlMark

yamlEndMark :: YamlMark

data Event

Constructors

EventStreamStart

EventStreamEnd

EventDocumentStart

EventDocumentEnd

EventAlias !AnchorName

EventScalar !ByteString !Tag !Style !Anchor

EventSequenceStart !Tag !SequenceStyle !Anchor

EventSequenceEnd

EventMappingStart !Tag !MappingStyle !Anchor

EventMappingEnd

data Style

Style for scalars - e.g. quoted / folded

Constructors

Any

Plain

SingleQuoted

DoubleQuoted

Literal

Folded

PlainNoTag

data SequenceStyle

Style for sequences - e.g. block or flow

Since: 0.9.0

Constructors

AnySequence

BlockSequence

FlowSequence

data MappingStyle

Style for mappings - e.g. block or flow

Since: 0.9.0

Constructors

AnyMapping

BlockMapping

FlowMapping

data Tag

Constructors

StrTag

FloatTag

NullTag

BoolTag

SetTag

IntTag

SeqTag

MapTag

UriTag String

NoTag

type AnchorName = String

type Anchor = Maybe AnchorName

Encoding and decoding

encode :: MonadResource m => ConduitM Event o m ByteString

encodeWith :: MonadResource m => FormatOptions -> ConduitM Event o m ByteString

Since: 0.10.2.0

decode :: MonadResource m => ByteString -> ConduitM i Event m ()

Create a conduit that yields events from a bytestring.

decodeMarked :: MonadResource m => ByteString -> ConduitM i MarkedEvent m ()

Create a conduit that yields marked events from a bytestring.

This conduit will yield identical events to that of "decode", but also includes start and end marks for each event.

Since: 0.10.4.0

encodeFile :: MonadResource m => FilePath -> ConduitM Event o m ()

decodeFile :: MonadResource m => FilePath -> ConduitM i Event m ()

Creata a conduit that yields events from a file.

decodeFileMarked :: MonadResource m => FilePath -> ConduitM i MarkedEvent m ()

Create a conduit that yields marked events from a file.

This conduit will yield identical events to that of "decodeFile", but also includes start and end marks for each event.

Since: 0.10.4.0

encodeFileWith :: MonadResource m => FormatOptions -> FilePath -> ConduitM Event o m ()

Since: 0.10.2.0

data FormatOptions

Contains options relating to the formatting (indendation, width) of the YAML output.

Since: 0.10.2.0

defaultFormatOptions :: FormatOptions

Since: 0.10.2.0

setWidth :: Maybe Int -> FormatOptions -> FormatOptions

Set the maximum number of columns in the YAML output, or Nothing for infinite. By default, the limit is 80 characters.

Since: 0.10.2.0

setTagRendering :: (Event -> TagRender) -> FormatOptions -> FormatOptions

Control when and whether tags are rendered to output.

Since: 0.1.1.0

renderScalarTags :: Event -> TagRender

A value for formatOptionsRenderTags that renders no collection tags but all scalar tags (unless suppressed with styles 'NoTag or PlainNoTag).

Since: 0.1.1.0

renderAllTags :: Event -> TagRender

A value for formatOptionsRenderTags that renders all tags (except NoTag tag and PlainNoTag style).

Since: 0.1.1.0

renderNoTags :: Event -> TagRender

A value for formatOptionsRenderTags that renders no tags.

Since: 0.1.1.0

renderUriTags :: Event -> TagRender

A value for formatOptionsRenderCollectionTags that renders tags which are instances of UriTag

Since: 0.1.1.0

Error handling

data YamlException

Constructors

YamlException String

YamlParseException
problem, context, index, position line, position column

Fields

yamlProblem :: String

yamlContext :: String

yamlProblemMark :: YamlMark

data YamlMark

The pointer position

Constructors

YamlMark

Fields

yamlIndex :: Int

yamlLine :: Int

yamlColumn :: Int